This was a college project where I developed a Supermarket List App, using the code provided from the teacher, the user is able to add the items and photos of the items in the supermarket list, as the price of the products.
The main objective of this project was to create a easy to use app, it will help the user to save money tracking their expenses, and sharing the app with their household.

I was looking to create an app that will facilitate the Social care professionals job, once they can use with their patients to improve their independency, allowing them to do their shop list on their own.